refatctor: reorganized the code structure
