--- .env.template
@@ -1,5 +1,0 @@
-# Qwen Image API Key Template
-# Copy this file to .env and replace the placeholder with your actual API key
-
-# Visit https://dashscope.console.aliyun.com/apiKey to get your API key
-DASHSCOPE_API_KEY=your_actual_api_key_here
