+++ config/.env.template
@@ -1,0 +1,9 @@
+# Qwen Image API Key Template
+# Copy this file to .env and replace the placeholder with your actual API key
+
+# Visit https://dashscope.console.aliyun.com/apiKey to get your API key
+# For international endpoint
+DASHSCOPE_API_KEY=your_international_api_key_here
+
+# For mainland China endpoint (optional, if different from international)
+DASHSCOPE_API_KEY_CHINA=your_china_api_key_here
