+++ core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Core package for Qwen Image API interactions
+"""
