+++ qwen_image/__init__.py
@@ -1,0 +1,3 @@
+"""
+Qwen Image package for ComfyUI
+"""
